% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpk-char_to_df.R
\name{char_to_df}
\alias{char_to_df}
\title{Convert From Hbase Char String to Dataframe}
\usage{
char_to_df(str)
}
\arguments{
\item{str}{The character string to be converted to a Dataframe.}
}
\value{
Dataframe containing IV curve data.
}
\description{
Hbase saves some data automatically as a long string in a column
of a dataframe. This function parses this string and creates
an additional dataframe from it.
}
\examples{
\donttest{
# generate Psuedo-IV Curves
df_full <- IVXbyX(df_wbw, corr_temp = "median", 4)

# subset a single row
degr_row <- df_full[1,]

degr_piv <- char_to_df(degr_row$piv)
}

}
