% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_wbw}
\alias{df_wbw}
\title{1 Year of raw outdoor IV curve data.}
\format{
Dataframe with 4140 objects in 10 variables:
\describe{
    \item{tmst}{A local PosixCT Timestamp}
    \item{ivdf}{IV Dataframe}
    \item{modt}{Module Temperature}
    \item{poa}{Plane of array (POA) irradiance}
    \item{isc}{Extracted Short Circuit Current by ddiv}
    \item{voc}{Extracted Open Current Voltage by ddiv}
    \item{rs}{Extracted Rs (series resistance) by ddiv}
    \item{pmp}{Extracted Maximum Power by ddiv}
    \item{imp}{Extracted Current at Maximum Power Point}
    \item{vmp}{Extracted Voltage at Maximum Power Point}
}
}
\usage{
df_wbw
}
\description{
This dataframe contains 1 year of anonymous raw outdoor time-series IV curve data.
}
\keyword{datasets}
