% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpk-IVcorr.R
\name{IVcorr}
\alias{IVcorr}
\title{Correct individual IV curve by 1-sun Voc and Isc}
\usage{
IVcorr(iv_row)
}
\arguments{
\item{iv_row}{A single row of psuedo-IV curve data.}
}
\value{
Returns a single row of corrected psuedo-IV curve data.
}
\description{
This method produces a dataframe with corrected
values that can be passed to power_loss_bat in order to calculate
power loss from IV curve data. This function is a component of \code{\link{IVcorr_full}},
which runs this function on an entire dataframe in a loop
}
