% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpk-IVcorr_full.R
\name{IVcorr_full}
\alias{IVcorr_full}
\title{Correct psuedo-IV curves by 1-sun Voc and Isc}
\usage{
IVcorr_full(df_red, df_piv)
}
\arguments{
\item{df_red}{A filtered dataframe containing IV curve data. Obtain with \code{\link{filter_df_raw}}}

\item{df_piv}{A dataframe of psuedo-IV curve data.
Obtain with \code{\link{read_df_raw}} or \code{\link{read_df_raw_from_csv}}.}
}
\value{
Returns A dataframe with corrected IV curves.
}
\description{
This method produces a dataframe with corrected IV curve
values that can be passed to power_loss_bat in order to calculate
power loss from IV curve data.
}
