% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpk-median_temp.R
\name{median_temp}
\alias{median_temp}
\title{Calculate Median Temperature}
\usage{
median_temp(df)
}
\arguments{
\item{df}{Dataframe containing IV Curve data (make sure it includes temperature).}
}
\value{
Returns an integer value of median reported temperature in the data.
}
\description{
This function calculates the median temperature throughout the data
to be used in corrections.
}
\examples{
T_corr <- median_temp(df_wbw)

}
