% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpk-char_to_df.R
\name{char_to_df}
\alias{char_to_df}
\title{Convert From Hbase Char String to Dataframe}
\usage{
char_to_df(str)
}
\arguments{
\item{str}{The character string to be converted to a Dataframe.}
}
\value{
Dataframe containing IV curve data.
}
\description{
This function parses the I-V curve string and creates
an additional dataframe with current and voltage columns from it.
}
\examples{
char_to_df(df_wbw$ivdf[1])

}
