% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpk-df2char.R
\name{df2chr}
\alias{df2chr}
\title{Convert From Dataframe to Hbase Char String}
\usage{
df2chr(df)
}
\arguments{
\item{df}{The dataframe, typically named "ivdf", to be converted.}
}
\value{
Returns a character string representing an IV curve.
}
\description{
The companion function to this one, \code{\link{char_to_df}},
parses this string and creates an additional dataframe from it.
This function works the other way, converting that dataframe back to a character
string.
}
\examples{
df2chr(char_to_df(df_wbw$ivdf[1]))

}
