% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpk-isc_1sun.R
\name{isc.1sun}
\alias{isc.1sun}
\title{Obtain 1-sun Isc Value}
\usage{
isc.1sun(isc, Irrad)
}
\arguments{
\item{isc}{Isc values}

\item{Irrad}{Irradiance values}
}
\value{
Returns a calculated Isc value at 1-sun Irradiance.
}
\description{
This function is used internally by IVxbyx and others for the
calculation of 1-sun Isc values based on Isc and Irradiance measurements.
}
\examples{
isc_1sun <- isc.1sun(df_wbw$isc, df_wbw$poa)

}
