% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpk-median_temp.R
\name{median_temp}
\alias{median_temp}
\title{Calculate Median Temperature}
\usage{
median_temp(df)
}
\arguments{
\item{df}{Dataframe containing timeseries irradiance (column name must be poa)
and module temperature (column name must be modt) in unit of Celsius.}
}
\value{
Returns an integer value of median reported module temperature of the data.
}
\description{
This function calculates the median module temperature throughout the data
to be used in corrections.
}
\examples{
T_corr <- median_temp(df_wbw)

}
