% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpk-read_data.R
\name{select_init_df}
\alias{select_init_df}
\title{Generate Initial Dataframe for power_loss_bat}
\usage{
select_init_df(df_raw, days = 21)
}
\arguments{
\item{df_raw}{The raw dataframe containing IV Curve measurement data.}

\item{days}{The number of initial days to subsample. Default: 21}
}
\value{
Subset of input dataframe.
}
\description{
The power_loss_bat function makes use of a dataframe containing
the first several days of psuedo-IV curves. This function creates
that dataframe for the user from the raw dataframe containing IV Curve
measurement data.
}
\examples{
df_init <- select_init_df(df_wbw, days = 21)

}
